requirements.txt
Appium-Python-Client==1.2.0
ImageHash==4.2.1
numpy==1.21.2
Pillow==8.3.2
pymediainfo==5.1.0
PyWavelets==1.1.1
PyYAML==5.4.1
scipy==1.7.1
selenium==3.141.0
six==1.16.0
urllib3==1.26.6
